[rename] #noissue : shared 폴더 위치 변경
