+++ src/pages/SearchFilePage.tsx
@@ -1,10 +1,10 @@
-import Navbar from "../../shared/ui/Navbar";
+import Navbar from "../shared/ui/Navbar";
 
 function SearchFilePage() {
   const navItems = [
     { path: "/", label: "파일 업로드" },
-    { path: "/searchfile", label: "파일 조회"},
-  ]
+    { path: "/searchfile", label: "파일 조회" },
+  ];
   return (
     <div>
       <Navbar items={navItems} />
