+++ src/pages/admin/AddUserPage.tsx
@@ -1,16 +1,16 @@
-import Navbar from "../../../shared/ui/Navbar";
+import Navbar from "../../shared/ui/Navbar";
 
 function AddUserPage() {
   const navItems = [
     { path: "/adminpage/workview", label: "업무 조회" },
     { path: "/adminpage/adduser", label: "사용자 등록" },
-  ]
+  ];
   return (
     <div>
-      <Navbar items={navItems}/>
+      <Navbar items={navItems} />
       adduser page
     </div>
-  )
+  );
 }
 
 export default AddUserPage;
