+++ src/shared/model/InputWithButton.type.ts
