+++ src/shared/ui/InputWithLabel.tsx
@@ -1,5 +1,5 @@
-import { Label } from "../../src/components/ui/label";
-import { Input } from "../../src/components/ui/input";
+import { Label } from "../../components/ui/label";
+import { Input } from "../../components/ui/input";
 import { InputWithLabelProps } from "../model";
 
 function InputWithLabel({ label, type, id, placeholder, value, onChange, success, error }: InputWithLabelProps) {
