+++ src/shared/ui/NavItem.tsx
