+++ src/shared/ui/Navbar.tsx
